(PC-24422)[PRO] refactor: simplify breadcrumb steps conditions
