+++ pro/src/components/IndividualOfferBreadcrumb/__specs__/IndividualOfferBreadcrumb.spec.tsx
@@ -14,7 +14,7 @@ import { individualOfferFactory } from 'utils/individualApiFactories'
 import { renderWithProviders } from 'utils/renderWithProviders'
 
 import { OFFER_WIZARD_STEP_IDS } from '../constants'
-import IndividualOfferBreadcrumb from '../IndividualOfferBreadcrumb'
+import { IndividualOfferBreadcrumb } from '../IndividualOfferBreadcrumb'
 
 const offerId = 12
 const stockId = 55
