--- pro/src/components/IndividualOfferBreadcrumb/index.ts
@@ -1,2 +1,0 @@
-export { default as IndividualOfferBreadcrumb } from './IndividualOfferBreadcrumb'
-export { OFFER_WIZARD_STEP_IDS } from './constants'
