+++ pro/src/core/Offers/utils/getIndividualOfferUrl/getIndividualOfferUrlV2/getIndividualOfferUrlV2.ts
@@ -1,6 +1,6 @@
 import { generatePath } from 'react-router-dom'
 
-import { OFFER_WIZARD_STEP_IDS } from 'components/IndividualOfferBreadcrumb'
+import { OFFER_WIZARD_STEP_IDS } from 'components/IndividualOfferBreadcrumb/constants'
 import { OFFER_WIZARD_MODE } from 'core/Offers/constants'
 
 interface GetIndividualOfferPathArgs {
