+++ pro/src/screens/IndividualOffer/ActionBar/__specs__/ActionBar.spec.tsx
@@ -2,7 +2,7 @@ import { screen } from '@testing-library/react'
 import { userEvent } from '@testing-library/user-event'
 import React from 'react'
 
-import { OFFER_WIZARD_STEP_IDS } from 'components/IndividualOfferBreadcrumb'
+import { OFFER_WIZARD_STEP_IDS } from 'components/IndividualOfferBreadcrumb/constants'
 import { renderWithProviders } from 'utils/renderWithProviders'
 
 import { ActionBar } from '..'
