+++ pro/src/screens/IndividualOffer/Informations/Informations.tsx
@@ -4,7 +4,7 @@ import { useLocation, useNavigate } from 'react-router-dom'
 
 import ConfirmDialog from 'components/Dialog/ConfirmDialog'
 import FormLayout from 'components/FormLayout'
-import { OFFER_WIZARD_STEP_IDS } from 'components/IndividualOfferBreadcrumb'
+import { OFFER_WIZARD_STEP_IDS } from 'components/IndividualOfferBreadcrumb/constants'
 import {
   IndividualOfferFormValues,
   IndividualOfferForm,
