+++ pro/src/screens/IndividualOffer/Informations/__specs__/InformationsScreen.creation.spec.tsx
@@ -7,7 +7,7 @@ import { api } from 'apiClient/api'
 import { ApiError, GetIndividualOfferResponseModel } from 'apiClient/v1'
 import { ApiRequestOptions } from 'apiClient/v1/core/ApiRequestOptions'
 import { ApiResult } from 'apiClient/v1/core/ApiResult'
-import { OFFER_WIZARD_STEP_IDS } from 'components/IndividualOfferBreadcrumb'
+import { OFFER_WIZARD_STEP_IDS } from 'components/IndividualOfferBreadcrumb/constants'
 import Notification from 'components/Notification/Notification'
 import {
   IndividualOfferContext,
