+++ pro/src/screens/IndividualOffer/Informations/__specs__/InformationsScreen.draft.spec.tsx
@@ -9,7 +9,7 @@ import {
   OfferStatus,
   WithdrawalTypeEnum,
 } from 'apiClient/v1'
-import { OFFER_WIZARD_STEP_IDS } from 'components/IndividualOfferBreadcrumb'
+import { OFFER_WIZARD_STEP_IDS } from 'components/IndividualOfferBreadcrumb/constants'
 import Notification from 'components/Notification/Notification'
 import {
   IndividualOfferContext,
