+++ pro/src/screens/IndividualOffer/Informations/utils/__specs__/computeNextStep.spec.ts
@@ -1,4 +1,4 @@
-import { OFFER_WIZARD_STEP_IDS } from 'components/IndividualOfferBreadcrumb'
+import { OFFER_WIZARD_STEP_IDS } from 'components/IndividualOfferBreadcrumb/constants'
 import { OFFER_WIZARD_MODE } from 'core/Offers/constants'
 
 import { computeNextStep } from '../computeNextStep'
