+++ pro/src/screens/IndividualOffer/StocksEventEdition/StocksEventEdition.tsx
@@ -8,7 +8,7 @@ import { api } from 'apiClient/api'
 import { PriceCategoryResponseModel } from 'apiClient/v1'
 import DialogBox from 'components/DialogBox'
 import FormLayout, { FormLayoutDescription } from 'components/FormLayout'
-import { OFFER_WIZARD_STEP_IDS } from 'components/IndividualOfferBreadcrumb'
+import { OFFER_WIZARD_STEP_IDS } from 'components/IndividualOfferBreadcrumb/constants'
 import { RouteLeavingGuardIndividualOffer } from 'components/RouteLeavingGuardIndividualOffer'
 import { StocksEvent } from 'components/StocksEventList/StocksEventList'
 import { useIndividualOfferContext } from 'context/IndividualOfferContext'
