+++ pro/src/screens/IndividualOffer/StocksEventEdition/__specs__/StocksEventEdition.RouteLeavingGuard.spec.tsx
@@ -9,7 +9,7 @@ import {
   GetIndividualOfferResponseModel,
   StockResponseModel,
 } from 'apiClient/v1'
-import { OFFER_WIZARD_STEP_IDS } from 'components/IndividualOfferBreadcrumb'
+import { OFFER_WIZARD_STEP_IDS } from 'components/IndividualOfferBreadcrumb/constants'
 import Notification from 'components/Notification/Notification'
 import {
   IndividualOfferContext,
