+++ pro/src/screens/IndividualOffer/StocksThing/__specs__/StocksThing.RouteLeavingGuard.spec.tsx
@@ -8,7 +8,7 @@ import {
   GetIndividualOfferResponseModel,
   StockResponseModel,
 } from 'apiClient/v1'
-import { OFFER_WIZARD_STEP_IDS } from 'components/IndividualOfferBreadcrumb'
+import { OFFER_WIZARD_STEP_IDS } from 'components/IndividualOfferBreadcrumb/constants'
 import {
   IndividualOfferContext,
   IndividualOfferContextValues,
