+++ pro/src/screens/IndividualOffer/Summary/OfferSection/OfferSection.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 
 import AccessibilitySummarySection from 'components/AccessibilitySummarySection'
-import { OFFER_WIZARD_STEP_IDS } from 'components/IndividualOfferBreadcrumb'
+import { OFFER_WIZARD_STEP_IDS } from 'components/IndividualOfferBreadcrumb/constants'
 import { SummaryLayout } from 'components/SummaryLayout'
 import { useIndividualOfferContext } from 'context/IndividualOfferContext'
 import {
