+++ pro/src/screens/IndividualOffer/Summary/PriceCategoriesSection/PriceCategoriesSection.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 
-import { OFFER_WIZARD_STEP_IDS } from 'components/IndividualOfferBreadcrumb'
+import { OFFER_WIZARD_STEP_IDS } from 'components/IndividualOfferBreadcrumb/constants'
 import { SummaryLayout } from 'components/SummaryLayout'
 import {
   Events,
