+++ pro/src/screens/IndividualOffer/Summary/StockSection/StockSection.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 
 import { OfferStatus } from 'apiClient/v1'
-import { OFFER_WIZARD_STEP_IDS } from 'components/IndividualOfferBreadcrumb'
+import { OFFER_WIZARD_STEP_IDS } from 'components/IndividualOfferBreadcrumb/constants'
 import { SummaryLayout } from 'components/SummaryLayout'
 import {
   Events,
