+++ pro/src/screens/IndividualOffer/Summary/StockSection/__specs__/StockSection.spec.tsx
@@ -4,7 +4,7 @@ import React from 'react'
 import { generatePath, Route, Routes } from 'react-router-dom'
 
 import { OfferStatus } from 'apiClient/v1'
-import { OFFER_WIZARD_STEP_IDS } from 'components/IndividualOfferBreadcrumb'
+import { OFFER_WIZARD_STEP_IDS } from 'components/IndividualOfferBreadcrumb/constants'
 import { OFFER_WIZARD_MODE } from 'core/Offers/constants'
 import { getIndividualOfferPath } from 'core/Offers/utils/getIndividualOfferUrl'
 import * as useAnalytics from 'hooks/useAnalytics'
