+++ pro/src/screens/IndividualOffer/Summary/Summary.tsx
@@ -3,7 +3,7 @@ import { useNavigate } from 'react-router-dom'
 
 import { BannerSummary } from 'components/Banner'
 import RedirectDialog from 'components/Dialog/RedirectDialog'
-import { OFFER_WIZARD_STEP_IDS } from 'components/IndividualOfferBreadcrumb'
+import { OFFER_WIZARD_STEP_IDS } from 'components/IndividualOfferBreadcrumb/constants'
 import { OfferAppPreview } from 'components/OfferAppPreview'
 import { SummaryLayout } from 'components/SummaryLayout'
 import { useIndividualOfferContext } from 'context/IndividualOfferContext'
