+++ pro/src/screens/IndividualOffer/utils/__specs__/logTo.spec.ts
@@ -1,4 +1,4 @@
-import { OFFER_WIZARD_STEP_IDS } from 'components/IndividualOfferBreadcrumb'
+import { OFFER_WIZARD_STEP_IDS } from 'components/IndividualOfferBreadcrumb/constants'
 
 import { logTo } from '../logTo'
